size() does not return an in, or a unsigned char, cast so the warning don't make baby jesus cry.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@7982 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/game/DynamicColor.cxx
@@ -67,7 +67,7 @@ void DynamicColorManager::update()
 int DynamicColorManager::addColor(DynamicColor* color)
 {
   colors.push_back (color);
-  return (colors.size() - 1);
+  return ((int)colors.size() - 1);
 }
 
 
@@ -84,7 +84,7 @@ DynamicColor* DynamicColorManager::getColor(int id)
 void * DynamicColorManager::pack(void *buf)
 {
   std::vector<DynamicColor*>::iterator it;
-  buf = nboPackUInt(buf, colors.size());
+  buf = nboPackUInt(buf, (int)colors.size());
   for (it = colors.begin(); it != colors.end(); it++) {
     DynamicColor* color = *it;
     buf = color->pack(buf);
